requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
h2o==3.22.1.1
numpy==1.15.2
pandas==0.23.4
prometheus-client==0.3.1
py==1.7.0
Pygments==2.2.0
requests==2.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
tornado==5.1.1
urllib3==1.23
pytest==4.1.1
pytest-forked==1.0.1
pytest-xdist==1.26.0
python-dateutil==2.7.3
